feat(route): Add `Sidebar` examples page & stories, resolving #53 #56 (#69)
* feat(route): Add `Sidebar` examples with new API

Adds the following examples for `Sidebar`s:

- [x] Default sidebar
- [x] Multi-level dropdown
- [x] Content separator
- [x] CTA button
- [x] Logo branding

* test(component): Add stories for `Sidebar`, resolving #56

The `Sidebar` is not highly configurable at the root
component, but there are a number of ways to set up
the contents, so we added those examples.

* refactor(story): Fix dupe story name in `Sidebar`

* fix(route): Restore accurate `Sidebar` example markup

Creating a component to reuse the code between
`storybook` and the page caused the page to display
the shortcut component instead of the actual
component's code.

* refactor(route): Use `flowbite-react` components in `Sidebar` example