upgrade dependencies to latest (#326)
* docs(component): fix incorrect `<Button>` colors

docs(component): fix incorrect `<Progress>` colors

* refactor(test): migrate tests to `@testing-library/react` `13.x`

Unfortunately, this is a huge diff because we are fundamentally changing every single test with this
migration. In this most recent version of RTL, we cut out a lot of the boilerplate that existed in
our old tests.

* fix(a11y): add accessible label to `Alert` close/dismiss button

* fix(component): prevent `<Accordion className="..">`

`className` was already omitted from props, so this isn't a breaking change.

Also prevents this for `<Avatar>`, `<Badge>`, and `<Breadcrumb>`

* fix(a11y): add accessible label to `<DarkThemeToggle>` icons

* feat(a11y): add `role="dialog"` to `<Modal>`s

See https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/dialog_role

* ci(coverage): allow PRs not meeting coverage threshold to pass CI

`codecov` provides a status check (https://docs.codecov.com/docs/commit-status) to PRs automatically
and will "fail" the PR if the threshold is not met. That prevents the PR from being merged, which is
undesirable.

* ci(coverage): increase coverage threshold to 5%

In other words, PRs can now be accepted with up to 5% of overall decreased code coverage compared to
`main`. In practice, coverage is not a hard and fast rule, more of a guide. And sometimes, you make
changes without tests. It just happens.