fix(/src/lib/components/Flowbite): fix `window is not defined` in `next.js` (#652)
Thanks @fsioni --

* fix(themecontext): fixing `ReferenceError: window is not defined error with Next.js

The variable `userPreferenceIsDark` was initialized with a value from the window object. This was causing an error with Next because of SSR. I added the check of the window, if it's on the server side, set to white

* fix(themecontext): fixing `ReferenceError: window is not defined` error with Next.js

The variable`userPreferenceIsDark` was initialized with a value from the window object. This was
causing an error with Next because of SSR. I added the check of the window, if it's on the server
side the value is set to `false`