docs: improve code preview (#798)
* docs(app/components/code-preview): add basic markup and styles

* docs(app/components/code-preview): improve box spacing, bg color

* docs(app/components/code-preview): add copy to clipboard

* docs(app/components/code-preview): add Github link template

* docs(app/components/code-preview): add typings for `react-copy-clipboard`

* docs(app/components/code-preview): add expand button

* docs(app/docs.css): style code preview text markup

* docs(app/docs.css): remove underlines from code previews

* docs(app/docs.css): fix custom CSS wasn't applied to code previews

We use a variety of custom CSS hooks in traditional CSS file (not Tailwind CSS) to override the
third-party plugin we use to render code previews. The `docs.css` file was not "cascading" as
intended in production, and all of the styles we wrote there suddenly disappeared in production
mode.

* docs(app/components/code-preview): remove pointless React package to copy to clipboard

This is rather easy to do without a third-party package!

* docs(app/components/code-preview): allow toggle dark mode on each code preview

This will only toggle the preview box itself - not the code. That reflects the behavior found on
flowbite.com.

* docs(app/components/code-preview): remove vert margin from code preview blocks

* docs(app/components/code-preview): add `Next.js`-style context to each preview

Provide full React context in the style of a `Next.js` page for each code preview, to make it more
immediately obviously how components should actually be used in the real world.

* docs(app/style.css): remove redundant Inter font

* docs(app/docs/components/*): add missing `import`s in component exmaples

Specifically, we're just doing the imports from `flowbite-react` - we
still need to add functionality to display an extra `import` line for
icons.

* build(fathom): use `<script defer>` to load instead of `npm` package

We are SSR and have access to the full HTML of the page via React 18, so we don't need something to
help us inject the script into the DOM.

* docs(app/components/navbar): fix Quickstart link shouldn't go to flowbite.com

* docs(app/docs/getting-started/introduction): replace nested `<p>`

* docs(app/docs/customize/theme): fix Option 3 code sample type

This should refer specifically to the subset `CustomFlowbiteTheme["button"]`.

* docs(app/components/code-preview): add functionality for Collapse code Button

* build(next.config): disable minified builds in production

We need to have access to the development-mode-only `displayName` of each imported FC, as well as to
the contents (bodies) of anonymous functions used inside component props in code examples. The only
way to do that is to disable Webpack minification.

* docs(app/docs/components/*): add missing external `import`s in code examples

* build(next.config): enable minification for prod with specific options

With these settings we reduce the bundle size back to where it was when we were minifying with
default settings, but still retain thedisplay names of FCs and the bodies of anonymous functions.

* docs(app/docs/components/*): update Edit on Github Button links

Provide a link to the Github.com page to view code of the primary component in each code preview.

* docs(app/docs/components/modal): input code previews manually

We need to input the code for Modal code previews manually, because our syntax highlighter resolves
the value of `<Modal show>`  instead of displaying the computation itself, and I can't see a way
around this behavior.

* docs(app/components/code-preview): fix code isn't highlighted after interaction

For components that you can interact with, like Pagination, doing so caused the code preview to stop
highlighting/coloring code.

* docs(code-preview): remove vertical scroll for code preview

* docs(cards): fix e-commerce card image path

* fix(theme): add cyan color to checkboxes and radios

* docs(sidebar): improve sidebar with CTA UI/UX

* docs(spinner): fix alignment spinner example imports

---------

Co-authored-by: Zoltán Szőgyényi <zoltan.szogyenyi@gmail.com>
Co-authored-by: Tanislav Robert <tanislav.robert@gmail.com>